ver2
Updated script after revision. The Poisson noise has been removed. Our intention behind applying Poisson noise, was not to reflect the biological distribution of specific variables, but rather to perturb the data even more, avoiding similar simulated patients. Acknowledging though that it was not the ideal practice, and it lacked medical realism, we decided to regenerate the simulated dataset, without including Poisson noise.